% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env.marine_1.0}
\alias{env.marine_1.0}
\title{Marine environments}
\format{
An object of class \code{data.frame} with 39 rows and 2 columns.
}
\usage{
data(env.marine_1.0)
}
\description{
Marine environments
}
\keyword{datasets}
