% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env.nonmarine_1.0}
\alias{env.nonmarine_1.0}
\title{Non-marine environments}
\format{
An object of class \code{data.frame} with 37 rows and 2 columns.
}
\usage{
data(env.nonmarine_1.0)
}
\description{
Non-marine environments
}
\keyword{datasets}
