% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fossilx}
\alias{fossilx-package}
\alias{fossilx}
\title{fossilx}
\description{
Extension and Post-processing of Fossil Occurrence Records
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/adamkocsis/fossilex/issues}
}

}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
