% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_main.R
\name{pbdb_extend}
\alias{pbdb_extend}
\title{Extend and post-process Paleobiology Database records}
\usage{
pbdb_extend(
  dat,
  include = list("tax.marine_1.0", "env.marine_1.0"),
  tax.level = c("genus"),
  tax.combine = "clgen",
  env.categories = "divDyn",
  strat = c("stg:1.0", "ten:1.0", "stb"),
  omit = list("lithification1:unlithified"),
  verbose = TRUE
)
}
\arguments{
\item{include}{Multiple element mean successive filtering, i.e. the intersection of sets. Character entries represent pre-defined datasets.}

\item{tax.level}{Taxonomic resolution. Currently only the genus level is defined. Arguments passed to pbdb_taxon_quality_filter.}

\item{tax.combine}{Combine columsn for homonomy evasion?}

\item{env.categories}{Should environment categories be defined?}

\item{strat}{Stratigraphic assignment definitions. Available are 'stg', 'ten' and 'stb', colons separate differnet versions.}

\item{omit}{What records should be omitted from the set?}
}
\description{
Declarative definition of post-processing steps of Paleobiology Database data
}
