% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_stratigraphy.R
\name{pbdb_strat_stb}
\alias{pbdb_strat_stb}
\title{Function to bin PBDB data to stb stages}
\usage{
pbdb_strat_stb(pbdb, ts, bin = "stb", stagecolumn = "time_contain")
}
\arguments{
\item{pbdb}{data.frame from database.}

\item{ts}{data.frame produced by StagesFromPBDB.}

\item{bin}{Character string, this will be the column name of the bin information}

\item{stagecolumn}{Charater string, of the time-bin information column in the PBDB}
}
\value{
A vector with integer bin numbers
}
\description{
Function to bin PBDB data to stb stages
}
\examples{
pbdb <- chronosphere::fetch("pbdb")
# new stages object
stages <- StagesFromPBDB(pbdb)
# bins from the pbdb output
pbdb$stb <- BinPBDB(pbdb, stages)
}
