% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_stratigraphy.R
\name{pbdb_strat_stg}
\alias{pbdb_strat_stg}
\title{Assign stage-level stratgiraphy values to PBDB occurrnces \code{stg}}
\usage{
pbdb_strat_stg(x, version = "1.0")
}
\arguments{
\item{x}{A Paleobiology Database occurrence data.frame (full output)}

\item{version}{A version number of the binning scheme.}
}
\value{
An occurrence data.frame with the column \code{stg} added to it.
}
\description{
This protocol uses the \code{stg} assignments of \code{divDyn}
}
\details{
The framework is described in detail
}
