% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_stratigraphy.R
\name{pbdb_strat_ten}
\alias{pbdb_strat_ten}
\title{Assign ten million year stratgiraphic bin values to PBDB occurrnces \code{stg}}
\usage{
pbdb_strat_ten(x, version = "1.0")
}
\arguments{
\item{x}{A Paleobiology Database occurrence data.frame (full output)}

\item{version}{A version number of the binning scheme.}
}
\value{
A vector of bin assignments.
}
\description{
This protocol uses the \code{ten} assignments of \code{divDyn}
}
\details{
The framework is described in detail
}
