% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_filter.R
\name{pbdb_taxon_quality_filter}
\alias{pbdb_taxon_quality_filter}
\title{Filter PBDB to a specific level of taxonomy}
\usage{
pbdb_taxon_quality_filter(x, level = "genus", omit = TRUE)
}
\arguments{
\item{x}{A PBDB data.frame.}

\item{level}{A taxonomic range}

\item{omit}{Should the filter omit occurrences, or create a new logical column that indicate which should be kept?}
}
\description{
The function ensures that
}
