% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_stratigraphy.R
\name{pbdb_timebins}
\alias{pbdb_timebins}
\title{Function to generate stage-lookup table from the Paleobiology Database occurrence data.frame}
\usage{
pbdb_timebins(
  pbdb,
  stagecolumn = "time_contain",
  min_ma = "min_ma",
  max_ma = "max_ma",
  bin = "stb"
)
}
\arguments{
\item{pbdb}{data.frame from database.}

\item{stagecolumn}{Charater string, of the time-bin information column}

\item{min_ma}{Character string, column name of the minimum ages}

\item{max_ma}{Character string, column name of the maximum ages}

\item{bin}{Character string, this will be the column name of the bin information}
}
\value{
A data.frame with the stages' name bottom, mid and top age and bin number.
}
\description{
Processess time-bin data downloaded from the database
}
