% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strat.camb_20180831}
\alias{strat.camb_20180831}
\title{Cambrian collection assignments}
\format{
An object of class \code{numeric} of length 5337.
}
\usage{
data(strat.camb_20180831)
}
\description{
Cambrian collection assignments
}
\keyword{datasets}
