% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strat.ord_20180831}
\alias{strat.ord_20180831}
\title{Ordovician collection assignments}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(strat.ord_20180831)
}
\description{
Ordovician collection assignments
}
\keyword{datasets}
