% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax.marine_1.0}
\alias{tax.marine_1.0}
\title{Marine taxa}
\format{
An object of class \code{data.frame} with 79 rows and 2 columns.
}
\usage{
data(tax.marine_1.0)
}
\description{
Marine taxa
}
\keyword{datasets}
